+++ photon-model-tasks/src/main/java/com/vmware/photon/controller/model/tasks/PhotonModelTaskServices.java
@@ -14,8 +14,10 @@
 package com.vmware.photon.controller.model.tasks;
 
 import com.vmware.photon.controller.model.tasks.monitoring.SingleResourceStatsCollectionTaskService;
+import com.vmware.photon.controller.model.tasks.monitoring.StatsAggregationTaskService;
 import com.vmware.photon.controller.model.tasks.monitoring.StatsCollectionTaskSchedulerService;
 import com.vmware.photon.controller.model.tasks.monitoring.StatsCollectionTaskService;
+
 import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.ServiceHost;
 
@@ -36,7 +38,8 @@ public class PhotonModelTaskServices {
             ProvisionFirewallTaskService.FACTORY_LINK,
             StatsCollectionTaskSchedulerService.FACTORY_LINK,
             StatsCollectionTaskService.FACTORY_LINK,
-            SingleResourceStatsCollectionTaskService.FACTORY_LINK };
+            SingleResourceStatsCollectionTaskService.FACTORY_LINK,
+            StatsAggregationTaskService.FACTORY_LINK,};
 
     public static void startServices(ServiceHost host) throws Throwable {
 
@@ -53,6 +56,7 @@ public class PhotonModelTaskServices {
         host.startFactory(new ProvisionFirewallTaskService());
         host.startFactory(new StatsCollectionTaskSchedulerService());
         host.startFactory(new StatsCollectionTaskService());
+        host.startFactory(new StatsAggregationTaskService());
         host.startFactory(new SingleResourceStatsCollectionTaskService());
     }
 }